Refacto jwt management for user service
No changes in logic, just renaming and moving code
