+++ backend/users/src/main/java/org/opfab/users/controllers/CurrentUserWithPerimetersController.java
@@ -12,7 +12,7 @@ package org.opfab.users.controllers;
 import org.opfab.useractiontracing.repositories.LastUserActionRepository;
 import org.opfab.useractiontracing.repositories.UserActionLogRepository;
 import org.opfab.useractiontracing.services.UserActionLogService;
-import org.opfab.users.configuration.oauth2.UserExtractor;
+import org.opfab.users.configuration.authorization.UserExtractor;
 import org.opfab.users.model.*;
 import org.opfab.utilities.eventbus.EventBus;
 import org.opfab.users.repositories.EntityRepository;
@@ -40,23 +40,24 @@ public class CurrentUserWithPerimetersController implements UserExtractor {
     private CurrentUserWithPerimetersService currentUserWithPerimetersService;
 
     public CurrentUserWithPerimetersController(UserRepository userRepository, GroupRepository groupRepository,
-                                               PerimeterRepository perimeterRepository, EntityRepository entityRepository,
-                                               UserSettingsRepository userSettingsRepository, UserActionLogRepository userActionLogRepository,
-                                               EventBus eventBus,
-                                               @Value("${operatorfabric.userActionLogActivated:true}") boolean userActionLogActivated,
-                                               @Value("${operatorfabric.users.getEmailFromUserInsteadOfSettings:false}") boolean getEmailFromUserInsteadOfSettings,
-                                               LastUserActionRepository lastUserActionRepository) {
+            PerimeterRepository perimeterRepository, EntityRepository entityRepository,
+            UserSettingsRepository userSettingsRepository, UserActionLogRepository userActionLogRepository,
+            EventBus eventBus,
+            @Value("${operatorfabric.userActionLogActivated:true}") boolean userActionLogActivated,
+            @Value("${operatorfabric.users.getEmailFromUserInsteadOfSettings:false}") boolean getEmailFromUserInsteadOfSettings,
+            LastUserActionRepository lastUserActionRepository) {
 
-            NotificationService notificationService = new NotificationService(userRepository, eventBus);
-            UsersService usersService = new UsersService(userRepository, groupRepository, entityRepository,
-                            perimeterRepository, notificationService);
-            UserActionLogService userActionLogService = new UserActionLogService(userActionLogRepository, lastUserActionRepository);
-            UserSettingsService userSettingsService = new UserSettingsService(userSettingsRepository, usersService,
-                            notificationService, userActionLogService, userActionLogActivated);
-            this.currentUserWithPerimetersService = new CurrentUserWithPerimetersService(usersService,
-                            userSettingsService,
-                            entityRepository,
-                            getEmailFromUserInsteadOfSettings);
+        NotificationService notificationService = new NotificationService(userRepository, eventBus);
+        UsersService usersService = new UsersService(userRepository, groupRepository, entityRepository,
+                perimeterRepository, notificationService);
+        UserActionLogService userActionLogService = new UserActionLogService(userActionLogRepository,
+                lastUserActionRepository);
+        UserSettingsService userSettingsService = new UserSettingsService(userSettingsRepository, usersService,
+                notificationService, userActionLogService, userActionLogActivated);
+        this.currentUserWithPerimetersService = new CurrentUserWithPerimetersService(usersService,
+                userSettingsService,
+                entityRepository,
+                getEmailFromUserInsteadOfSettings);
     }
 
     @GetMapping(produces = { "application/json" })
